Flatten hierarchies by removing Parent Notation
The dataset search is currently unaware of concept-schemes or their hierarchies, presenting only the `skos:hasTopConcept` codes.

This means that some dimension filters show only a choice of e.g. "Total" which is neither informative nor practical.

The ultimate solution is to make the dataset search concept-scheme aware.

In the mean-time this comment removes the hierarchy, flattening these schemes.

Note that several larger schemes (particularly for the product dimension) retain their hierarchy as they would simply be unmanageable otherwise.

Future committers will likely want a combination of revert and rebase to reverse this patch cleanly in future.